hopefully fix auth on prod build
